Add default wallets for WalletConnect Modal (#296)
WalletConnect Modal SDK needs the init() to be called as soon as the app
launches, so we can't wait until the wallet IDs from env.json is loaded.
We will store the wallet IDs from env.json to local storage, and load
them at app launch. For the first app launch, we will just use a default
list.